requirements.txt
# =============================================================================
# AI-Powered Employee Helpdesk Agent - Dependencies
# =============================================================================

# Core Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1

# Microsoft Bot Framework
botbuilder-core==4.16.1
botbuilder-integration-aiohttp==4.16.1
botbuilder-dialogs==4.16.1
botbuilder-schema==4.16.1
aiohttp==3.9.3

# LangChain
langchain==0.1.9
langchain-openai==0.0.6
langchain-core==0.1.27
langgraph==0.0.26

# MCP (Model Context Protocol)
mcp==1.0.0

# OpenAI
openai==1.12.0
tiktoken==0.6.0

# Azure Services
azure-identity==1.15.0
azure-search-documents==11.4.0
azure-core==1.30.0
msal==1.26.0

# HTTP Client
httpx==0.26.0
aiofiles==23.2.1

# Document Processing
pypdf==4.0.2
python-docx==1.1.0
unstructured==0.12.4

# Utilities
tenacity==8.2.3
structlog==24.1.0
python-dateutil==2.8.2

# Testing
pytest==8.0.1
pytest-asyncio==0.23.4
pytest-cov==4.1.0
respx==0.20.2
